Fix checkboxes appearance
I've done this by duplicating css rules from nhs
with wider selectors that accommodate the extra hidden field added by
dotnet. Not clear why it wasn't adding it earlier, but it's desired
behaviour and the other option was to not use `asp-for` which felt like
a waste.

Also moved all stylesheets to scss while there.