NTBS-456 Add hidden save button
Moving the action buttons to the top of the page made them the primary button for the form, which browsers trigger on hitting 'Enter'.
To retain the behaviour where main button is still the one being triggered on 'Enter', we need a hidden copy of it at the top of the form.