Set deployment definition images away from `latest`
This change is done in response to the scenario of calling `kubectl apply uat.yml` directly,
 rather than through the release script (which follows it up with `kubectl set image ...`). It had the following adverse effects:
 - the newest version of code got released to uat
 - from that point onwards uat auto-updated to the newest image, until the release script got used again

 This was caused by the fact that:
 - `ntbscontainerregistry.azurecr.io/ntbs-service` implicitly described the tag `ntbs-service:latest`
 - we have `imagePullPolicy: Always` set on the definitions (and it is also the default behaviour for `latest` tags)

The change in this commit should stop this behaviour by giving each environment a non-existent image to pull in as the baseline. That way if `apply` is called, the new replica set would simply fail to start and the old one will keep running.
It will not affect the current deployment process (which, as mentioned above, overrides the image value anyways).