NTBS-1047 Set SameSite mode for AddWsFederation
This is an attempt at fixing infinite-loop login problems experienced on chrome recently.
The assumption is that Chrome's recent enforcement of more strict cookie policies (https://web.dev/samesite-cookies-explained/)
is causing the correlation cookie to be rejected by Chrome, so when the user is logged in, NTBS has no way of recognizing it and requests another login.