NTBS-1064 Display enhanced surveillance pages always if there is any data in them
Situations where a notification was not marked for enhanced surveillance by usual means (ie. mostly through matched test results) but *does* contain information entered in the enhanced surveillance forms could be treated as theoretical corner cases.
But in practice it's not too hard to create them, especially during the migration process. Although they might (and should) be sparse in the final release, they have been quite common during development.
Finally, even if corner-casey, we should never *hide away* information that was entered into the system.
Due to all of the above, it made most sense to surface the information whenever it exists, regardless of what the indicators say.