NTBS-1530 Move from ingresses to routes
On openshift, ingress is basically treated as way to create a route:
https://www.openshift.com/blog/kubernetes-ingress-vs-openshift-route

Sam and I were not able to find a way to define an ingress with tls edge-termination which would use the default certificate (there doesn't seem to be a way around passing a certificate secret to an ingress).
Routes have that capability, and as the simplest way to fulfill the need, they won that battle.