NTBS-1588 Add NO LOCK clauses on import reads
These `SELECT` statements are reading data that cannot change during the runtime of the app (as it's in the static migration db). Having the no lock clauses might help speed the migration up somewhat.

Reluctant to put the same hint on the main queries in the file, as they check for whether these items are marked as migrated already - which will change dynamically.