Update display name formatting
Add a display name formatting helper to remove long trust names from
display names retrieved from Azure AD. For example:
"SMITH, John (Some NHS trust name)" would format to "SMITH, John".

Add some unit tests for this helper.

Use this when formatting the user name in the website header, and also
when populating the User table in the regular sync job.

Add a fallback to use given and family names as a display name if there
is no display name present (which is the case for some legacy users).