NTBS-2082 and NTBS-2085 field accessibility fixes
Form tweaks to improve accessibility: making sure every form input has
a label, and that subgroups of inputs are in fieldsets, and all
fieldsets have a legend.

This led to some styling tweaks for the search page, because the dates
and sex input groups needed to be put in a form field, but making a
fieldset legend appear inline with its inputs is not as trivial as you
would hope (in particular, flex boxes and form legends do not seem to
play nicely together).