Fix EF ownership issue for drug resistance job
Fixed an issue introduced when upgrading to .NET 5 (well the change was
made in v2 -> v3), where owned entities cannot be queried without the
owner using a tracking query. This has been causing some errors on
Sentry, but should now be fixed.

Fix by adding a notification to the NotificationForDrugResistanceImport
class, instead of just its ID, because this object is owned by the
Notification object.