Add automatic MDR alert dismissal
Dismiss a notification's MDR alert when the user fills in its MDR
questionnaire. Do this by running
EnhancedSurveillanceAlertsService.CreateOrDismissMdrAlert after
editting the MDR questionnaire.

Modify the CreateOrDismissAlert service method to check whether the
questionnaire has been filled out when deciding whether to create or
dismiss the alert. To do this, also add MDR data to the
NotificationForDrugResistanceImport projection. This adds in a few
fields but hopefully will not have a significant impact on the
performance of the DrugResistanceProfileUpdateJob.

Amend the EnhancedSurveillanceAlertsService tests to cover this new
logic.