Turn off authentication cookie sliding expiration
Sliding cookie expiration means that an authentication cookie halfway
through its life will get replaced automatically when you make a
request to the server with this cookie.

This allows an attacker to maintain an authentication cookie forever
if they were to steal one. Because of this, we need to turn this
feature off.