Reduce the likelihood of a race in Selenium tests
Add an implicit wait timespan to Selenium tests; which tells Selenium
to wait some time before claiming an element is not on the page. This
gives some time for JS to modify the page from the previous step,
making JS much more likely to win its race against Selenium.

I was having issues with the "Fill in M. bovis occupation exposure
page" scenario even though no relevant changes had happened recently.
The issue was as described above, Selenium was not giving JS enough
chance to add the "add-new-button" to the page before failing the
test, but they now all work.