Add unknowns to M. bovis details table
For the M. bovis exposure summaries HasExposureToKnownCases,
HasUnpasteurisedMilkConsumption, HasOccupationExposure, and
HasAnimalExposure in MBovisDetails, switch from storing these as bits
and instead use the Status enum (Yes, No, Unknown), in effect adding an
"unknown" option to each of these questions.

Add this additional option to the website front-end, and amend the
relevant tests and EF context.

Match current behaviour with legacy migration for M. bovis details, but
this will be improved soon using this change in NTBS-2170.