Improve M. bovis questionnaire migration
If no questionnaire forms have been filled out in the legacy system,
leave all the M. bovis question responses as NULL.

If some of the M. bovis questionnaire has been filled out, fill in the
M. bovis details questions as either Yes or Unknown, instead of Yes or
No. This is because an absence of data in the migration database might
because the answer to that exposure question was not known, or it was
known to be "no". Because we can't tell the difference, we fill the
values in as "unknown".