Fix search button for narrow screens
The search button style for narrow screens is different to what we
need, it is larger and green. We are not really using the button in
the context that is written for.

Instead, make our own search button style based on the NHS search
button style, without the unwanted style change for narrow screens.