Split transfer events into own period
Visually separate transfer out events from the other treatment events
in their period. To this end, make a TreatmentPeriod class that
contains a list of events, corresponding to each period on the website.
These TreatmentPeriod objects are numbered, and a flag to denote a
transfer-out-only period, to be displayed separately but without a
number.

Amend unit tests to reflect the new way treatment episodes are grouped.