Remove Sentry install hotfix
For a while the @sentry/cli install script was not being run, leading
to npm run build:prod failing. This issue seems to have been fixed
after the latest round of package updates, so we can remove the Docker
build step that runs this install script that ought to have run.