Refactor reporting page and its redirects
Use a more basic reporting page, redirecting to the reporting service
page if it exists, and giving an error message if not. Change the nav
bar to open the reporting service page in a new tab directly, instead
of using the /Reports route. Simplify the reports routing, removing the
possibility of individually indexed reports.

Provide the reporting service page URL using a new
ReportingLinksService which reads from the ExternalLinks configuration.

Rename the ExternalLinks:ReportingUri to the more consistent
ExternalLinks:ReportingOverview