Get cluster match changes from specimen DB
The cluster matches are now stored in the specimen matching DB, not
reporting, so get the changes from there.

Change how the Cluster Data Quality alert is generated: don't generate
it using the DataQualityAlerts job, instead create or dismiss the alert
when changes to the cluster ID are being applied by the
NotificationClusterUpdate job.

Amend the NotificationService tests to cover this create or dismiss
behaviour.