Add tests for the notification creation process
Add integration tests for the redirect behaviour of the Create route,
and check that the drug resistance profile has been set to a default
value (this is done by the database, configured in NtbsContext, so we
need an integration test for that).

Add a unit test to check that new notifications are created in the
notifying user's TB service, and has their case manager ID, so that
they are able to edit their newly created notification draft.