Dismiss alerts if notification is closed
Don't open data quality alerts for notifications that are closed, and
when a notification is being closed all of the alerts for it should
also be closed.

The only exception to this is that the duplicate notification alert
should only be closed if both notifications are closed.

To do this, run a new method in the alert service to close all alerts
for a notification, from the notification closing logic.

Add some tests for this AlertService logic.