Describe environment in phase banner
Modify the phase banner so it only displays when live data should not
be in the system, and have it describe the environment (eg. "Training"
or "Development").

Adjust the phase banner styles to make it more eye-catching, and move
it to below the app name in accordance with GOV.UK design system
recommendations.