Make select list values consistent
Fix a bug where the manual test result, result value would be lost from
the edit page, because the select list used was not of the format that
the Vue Filtered-Dropdown component was being told by the page.

The approach used by treatment event outcome subtype was correct;
text as the enum name and the value as the enum integer. Extract out
this calculation to OptionValue.FromEnum, and call it from both places.