Use Graph Client pagination during UserSync
We were having issues running UserSync because there were too many
different groups under the BaseUserGroup. This is because the request
returns the results in pages of 100 results, and there were ~250
results, so ~60% of the groups were being ignored.

By using NextPageRequest we can read all of the pages in sequence.

Also, tidy up some of the graph client queries, such as the base group
query to use an equality check on displayname instead of a weird prefix
and filter search.