Fix a specimen import result bug
When failing to import a specimen, this is noted in the ImportResult
object. However, because that object will have recently been marked as
successful, previously denoted by a null error list for that
notification, then adding this error to the null list will cause a
NullPointerException. Instead, we now encode success as an empty list
of errors, instead of a null list of errors, so this list can be added
to after the notification has been imported to track warnings.