Refactor AD services
Tweak some types in the interfaces of various AD Directory services.

Heavily refactor the Azure AD service to tidy it up, and to stop
suppressing exceptions. Change the way responses are read to only send
valid requests (eg. to stop trying to get Group members of objects that
aren't groups, like DirectoryRoles), so we don't regularly throw
exceptions that we then need to suppress.