Import LTBR Patient ID from new migration field
LTBR Patient ID is now being tracked separately from group ID, so
use this new field in the places GroupID was being used as an alias for
LTBR Patient ID.

When searching for legacy notifications, search using LTBR patient ID,
as well as ETS and LTBR notification IDs.

When importing a legacy notification, only do it by notification ID.
This means that if an ETS notification and an LTBR patient ID are the
same (for different patients), importing by ETS ID will not import that
LTBR notification, when it wasn't really wanted.