Stop legacy hospitals appearing on edit page
Legacy hospitals should only be shown on the edit page if it is that
notification's hospital, otherwise they need to be active. The
filtered-dropdown component was getting a list of hospitals containing
legacy hospitals.

Make the logic for getting a list of hospitals for the edit page common
across the two places it is called, and include the correct filtering
steps.