Use projection for notification banner
Project the info needed for a notification banner out of the request
for it, and the build the model separately.

This is only used for the search page at the moment, but that is the
page with the most notification banners by far, as many as 50, compared
to a linked notification page which might have a handful.