Refactor filters for treatment outcome alerts
Filter by notification status as well as date range for the first pass
of filtering for treatment outcome DQ alerts. The more complex
treatment outcome analysis is done in memory for each possibly eligible
notification, so the more precisely we define that set the less
notifications the job has to consider each night.