Fix validation bug on lab results page
When fetching a notification by ID for a manual match, if it is null
or empty just clear the notification details, and notification ID
validation errors, instead of 500'ing. Indicate this by returning an
empty object instead of one with an errorMessages field.

Add some logic to the FetchSpecimenPotentialMatch component to
facilitate this.

NTBS-2573