Fix manage notifications button styling
Make the focus yellow fill the whole button, instead of just around the
text. Fix some issues with borders, especially the interaction between
round and sharp corners between the button and the box it expands. Fix
an issue with a plus being shown instead of a minus when the box is
expanded but not in focus.

NTBS-1280