Add a PostLogout page
Add a page telling the user the have been logged out, with a button to
log back in. This page can be accessed anonymously (by users who have
not been authenticated). Before we would redirect to a log in page
with a correlation cookie that would time out after 15 minutes, so the
user would need to log in twice (with the first causing a warning in
Sentry). Now, by redirecting to this post-logout page the user can
choose when they get sent to Azure AD with that cookie.

This won't prevent the correlation cookie issue, if the user visits
the log in page and leaves it for 15 mins it will still happen, but
now the user needs to do something to visit the login page, instead of
getting redirected after timing out.

NTBS-2666